requirements.txt
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
blinker==1.8.2
certifi==2024.2.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
dnspython==2.6.1
Flask==3.0.3
flask-restx==1.3.0
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
idna==3.6
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
numpy==2.0.0
openai==1.12.0
pandas==2.2.2
pydantic==2.6.1
pydantic_core==2.16.2
pymongo==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
referencing==0.35.1
rpds-py==0.19.0
six==1.16.0
sniffio==1.3.0
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2024.1
Werkzeug==3.0.3
